+++ pennylane_lightning/core/simulators/lightning_gpu/catalyst/tests/Test_LightningGPUSimulator.cpp
@@ -763,6 +763,29 @@ TEST_CASE("LightningGPUSimulator::GateSet", "[GateSet]") {
             Catch::Contains("Controlled wires/values size mismatch"));
     }
 
+    SECTION("Controlled GlobalPhase") {
+        std::unique_ptr<LGPUSimulator> LGPUsim =
+            std::make_unique<LGPUSimulator>();
+
+        constexpr std::size_t n_qubits = 4;
+        std::vector<intptr_t> Qs;
+        Qs.reserve(n_qubits);
+
+        for (std::size_t i = 0; i < n_qubits; i++) {
+            Qs[i] = LGPUsim->AllocateQubit();
+        }
+
+        LGPUsim->NamedOperation("GlobalPhase", {M_PI}, {}, false);
+        LGPUsim->NamedOperation("GlobalPhase", {M_PI_2}, {}, true);
+        LGPUsim->NamedOperation("GlobalPhase", {M_PI_2}, {}, false, {Qs[3]},
+                                {true});
+        LGPUsim->NamedOperation("GlobalPhase", {M_PI_4}, {}, false,
+                                {Qs[1], Qs[0]}, {true, false});
+
+        ObsType pz = LGPUSim->NamedObservable("PauliZ", {Qs[0]});
+        CHECK(LGPUSim->Expval(pz) == Approx(1.0).margin(1e-5));
+    }
+
     SECTION("Test setStateVector") {
         std::unique_ptr<LGPUSimulator> LGPUsim =
             std::make_unique<LGPUSimulator>();